Reorg map loader to director
