+++ src/PlayerUpdateTask.hpp
@@ -1,7 +1,7 @@
 /******************************************************************************
 
     GPLv3 License
-    Copyright (c) 2022 Aria Janke
+    Copyright (c) 2023 Aria Janke
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -20,47 +20,7 @@
 
 #pragma once
 
-#include "../Defs.hpp"
-#include "../Components.hpp"
-#include "TiledMapLoader.hpp"
-
-#include "MapRegionTracker.hpp"
-
-namespace point_and_plane {
-    class Driver;
-} // end of point_and_plane namespace
-
-/** @brief The MapLoadingDirector is responsible for loading map segments.
- *
- *  Map segments are loaded depending on the player's state.
- */
-class MapLoadingDirector final {
-public:
-    using PpDriver = point_and_plane::Driver;
-
-    MapLoadingDirector(PpDriver * ppdriver, Size2I chunk_size):
-        m_ppdriver(ppdriver),
-        m_chunk_size(chunk_size)
-    {}
-
-    SharedPtr<BackgroundTask> begin_initial_map_loading
-        (const char * initial_map, Platform & platform, const Entity & player_physics);
-
-    void on_every_frame(TaskCallbacks & callbacks, const Entity & physics_ent);
-
-private:
-    static Vector2I to_segment_location
-        (const Vector & location, const Size2I & segment_size);
-
-    void check_for_other_map_segments
-        (TaskCallbacks & callbacks, const Entity & physics_ent);
-
-    // there's only one per game and it never changes
-    PpDriver * m_ppdriver = nullptr;
-    Size2I m_chunk_size;
-    std::vector<TiledMapLoader> m_active_loaders;
-    MapRegionTracker m_region_tracker;
-};
+#include "map-director/MapDirector.hpp"
 
 /// all things the player needs to do everyframe
 ///
@@ -68,7 +28,7 @@ private:
 class PlayerUpdateTask final : public EveryFrameTask {
 public:
     PlayerUpdateTask
-        (MapLoadingDirector && map_director, const EntityRef & physics_ent):
+        (SharedPtr<MapDirector_> && map_director, const EntityRef & physics_ent):
         m_map_director(std::move(map_director)),
         m_physics_ent(physics_ent)
     {}
@@ -80,7 +40,7 @@ public:
 private:
     static void check_fall_below(Entity & ent);
 
-    MapLoadingDirector m_map_director;
+    SharedPtr<MapDirector_> m_map_director;
     // | extremely important that the task is *not* owning
     // v the reason entity refs exists
     EntityRef m_physics_ent;
