+++ src/map-director/GidTidTranslator.cpp
