+++ src/map-director/GidTidTranslator.hpp
