+++ src/map-director/MapRegion.cpp
