+++ src/map-director/MapRegionTracker.cpp
