+++ src/map-director/MapRegionTracker.hpp
