+++ src/map-director/ParseHelpers.cpp
