+++ src/map-director/ParseHelpers.hpp
