+++ src/map-director/ProducableTileFiller.cpp
@@ -1,0 +1,36 @@
+/******************************************************************************
+
+    GPLv3 License
+    Copyright (c) 2023 Aria Janke
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+*****************************************************************************/
+#if 0
+#include "ProducableTileFiller.hpp"
+#include "slopes-group-filler.hpp"
+
+namespace {
+
+
+} // end of <anonymous> namespace
+
+/* static */ SharedPtr<ProducableTileFiller> ProducableTileFiller::make_slopes_group_filler
+    (const TileSetXmlGrid & xml_grid, Platform & platform)
+{
+    auto rv = make_shared<SlopeGroupFiller>();
+    rv->load(xml_grid, platform);
+    return rv;
+}
+#endif
