+++ src/map-director/TileFactory.cpp
