+++ src/map-director/TileSetPropertiesGrid.cpp
