+++ src/map-director/TileSetPropertiesGrid.hpp
