+++ src/map-director/TileTexture.cpp
