+++ src/map-director/TileTexture.hpp
