+++ src/map-director/ViewGrid.hpp
