+++ src/map-director/map-loader-helpers.cpp
