+++ src/map-director/map-loader-helpers.hpp
