+++ src/map-director/map-loader.hpp
