+++ src/map-director/slopes-group-filler/WallTileFactoryBase.hpp
